+++ CHANGELOG.md
@@ -1,4 +1,6 @@
-## version 0.1.0 flutter Chat  android and iOS 
-## version 0.1.1 add screen widgets  
-## version 0.1.2 example added
+## version 0.1.4 dependencies upgrade
 ## version 0.1.3 widget seprated
+## version 0.1.2 example added
+## version 0.1.1 add screen widgets
+## version 0.1.0 flutter Chat  android and iOS 
+